Enable caching for 3 expressions.
Caching is useful for when we have duplicate sub-trees in the
expression. For example,

    auto a = b + c;
    auto d = trans(a);
    auto e = a * d;

The '+' operation for b and c needs to be evaluated twice. With
caching, however, it only needs to be evaluated once.

Several improvements are necessary, though, for caching to work
efficiently. First of all, we don't need caching when we don't
have duplicate parts in the graph. We need some way to detect
where caching is needed and where it isn't needed. Second,
it needs to mix well with simplifications (optimizations). This
needs to be investigated further.